Extract thumb fetch into AlbumsDataFetcher class
